requirements.txt
langchain
langgraph
graphviz
IPython
langchain-openai
python-dotenv 